refactor(simulation): move the price level snapshot tracker to its own file
