+++ crates/tycho-simulation/src/price_level_stream/mod.rs
@@ -41,4 +41,7 @@ pub mod fallback_router;
 pub mod state;
 pub mod stream;
 mod telemetry;
+#[cfg(test)]
+mod test_support;
 mod titan;
+mod tracker;
