+++ crates/tycho-simulation/src/price_level_stream/stream.rs
@@ -1,16 +1,11 @@
 use std::{
-    collections::{hash_map::Entry, HashMap, HashSet},
+    collections::{HashMap, HashSet},
     time::Duration,
 };
 
-use chrono::Utc;
 use num_bigint::BigUint;
 use tokio_stream::{Stream, StreamExt};
-use tycho_common::{
-    models::{token::Token, Chain},
-    simulation::protocol_sim::ProtocolSim,
-    Bytes,
-};
+use tycho_common::{models::token::Token, Bytes};
 
 use super::{
     config::{
@@ -18,13 +13,10 @@ use super::{
         DEFAULT_AUTO_DETECTED_GAS_COST,
     },
     fallback_router::fetch_fallback_router_venues,
-    state::{PriceLevelStreamQuote, PriceLevelStreamState},
-    titan::{
-        self, ConnectionSettings, TitanPairLevels, TitanPammLevels, TitanPriceLevel,
-        TitanPriceLevelMessage, TITAN_PRICE_LEVEL_URL,
-    },
+    titan::{self, ConnectionSettings, TITAN_PRICE_LEVEL_URL},
+    tracker::SnapshotTracker,
 };
-use crate::protocol::models::{ProtocolComponent, Update};
+use crate::protocol::models::Update;
 
 /// Static attribute under which each emitted component carries its pAMM venue address.
 pub const PAMM_ADDRESS_ATTRIBUTE: &str = "pamm_address";
@@ -312,454 +304,14 @@ async fn fetch_router_venues(rpc_url: Option<String>) -> HashSet<Bytes> {
     }
 }
 
-/// Turns Titan frames into [`Update`]s, tracking the previously emitted components so pair
-/// additions and removals can be diffed against the last snapshot.
-struct SnapshotTracker {
-    registry: HashMap<Bytes, PriceLevelStreamConfig>,
-    /// Venues excluded from auto-detection. The builder keeps this disjoint from the registry:
-    /// denying removes any registration and registering removes any denial.
-    denied: HashSet<Bytes>,
-    tokens: HashMap<Bytes, Token>,
-    /// Whether frames from pAMMs absent from the registry get an address-named configuration
-    /// synthesized (and cached in the registry) instead of being skipped.
-    auto_detect: bool,
-    /// The per-swap gas cost synthesized auto-detected configurations are served with.
-    auto_detected_gas_cost: BigUint,
-    /// Venues whose components are emitted under the `propammfallback:` family, so their swaps
-    /// execute through Titan's PropAMMRouter instead of the venue directly.
-    router_venues: HashSet<Bytes>,
-    /// Components of the last emitted snapshot, across all pAMMs. A frame is a complete
-    /// snapshot of everything Titan currently streams, so removals are diffed globally: a
-    /// known component a frame does not re-emit is gone — including when its venue vanishes
-    /// from the stream entirely.
-    components: HashMap<String, ProtocolComponent>,
-    /// The newest block number processed so far. Frames targeting an older block (e.g.
-    /// delivered around a reconnect) are stale and skipped wholesale — processing one would
-    /// emit superseded states and churn the global diff.
-    newest_block: u64,
-}
-
-impl SnapshotTracker {
-    fn new(
-        registry: HashMap<Bytes, PriceLevelStreamConfig>,
-        denied: HashSet<Bytes>,
-        tokens: HashMap<Bytes, Token>,
-        auto_detect: bool,
-        auto_detected_gas_cost: BigUint,
-        router_venues: HashSet<Bytes>,
-    ) -> Self {
-        Self {
-            registry,
-            denied,
-            tokens,
-            auto_detect,
-            auto_detected_gas_cost,
-            router_venues,
-            components: HashMap::new(),
-            newest_block: 0,
-        }
-    }
-
-    /// Processes one frame into an [`Update`], or `None` if the frame targets an older block
-    /// than an already processed one or contains nothing relevant (no registered pAMM with at
-    /// least one known pair or a pair removal).
-    fn process(&mut self, message: TitanPriceLevelMessage) -> Option<Update> {
-        if message.block_number < self.newest_block {
-            tracing::warn!(
-                block_number = message.block_number,
-                newest_block = self.newest_block,
-                "Skipping out-of-order price level frame"
-            );
-            return None;
-        }
-        self.newest_block = message.block_number;
-
-        let mut states: HashMap<String, Box<dyn ProtocolSim>> = HashMap::new();
-        let mut new_pairs = HashMap::new();
-        // The frame is a complete snapshot: every known component is presumed gone until the
-        // frame re-emits it below.
-        let mut previous = std::mem::take(&mut self.components);
-
-        for TitanPammLevels { pamm, pairs } in message.pamms {
-            let config = match self.registry.entry(pamm.clone()) {
-                Entry::Occupied(entry) => &*entry.into_mut(),
-                Entry::Vacant(entry) => {
-                    if !self.auto_detect {
-                        tracing::debug!(%pamm, "Skipping unregistered pAMM");
-                        continue;
-                    }
-                    if self.denied.contains(&pamm) {
-                        tracing::debug!(%pamm, "Skipping denied pAMM");
-                        continue;
-                    }
-                    tracing::info!(%pamm, "Serving auto-detected pAMM");
-                    &*entry.insert(PriceLevelStreamConfig::auto_detected(
-                        pamm.clone(),
-                        self.auto_detected_gas_cost.clone(),
-                    ))
-                }
-            };
-
-            // Merge the frame's per-direction ladders into one entry per unordered token pair.
-            let mut merged_pairs: HashMap<(Bytes, Bytes), (Vec<_>, Vec<_>)> = HashMap::new();
-            for TitanPairLevels { token_in, token_out, order_book } in pairs {
-                if !self.tokens.contains_key(&token_in) || !self.tokens.contains_key(&token_out) {
-                    tracing::debug!(%token_in, %token_out, "Skipping pair with unknown token");
-                    continue;
-                }
-                let sells_token0 = token_in < token_out;
-                let key = if sells_token0 {
-                    (token_in.clone(), token_out.clone())
-                } else {
-                    (token_out.clone(), token_in.clone())
-                };
-                let quotes = order_book
-                    .into_iter()
-                    .map(|TitanPriceLevel { amount_in, amount_out }| {
-                        PriceLevelStreamQuote::new(amount_in, amount_out)
-                    })
-                    .collect();
-                let entry = merged_pairs.entry(key).or_default();
-                if sells_token0 {
-                    entry.0 = quotes;
-                } else {
-                    entry.1 = quotes;
-                }
-            }
-
-            for ((token0, token1), (quotes_0_to_1, quotes_1_to_0)) in merged_pairs {
-                let id = component_id(&config.address, &token0, &token1);
-                let id_string = id.to_string();
-                let component = previous
-                    .remove(&id_string)
-                    .unwrap_or_else(|| {
-                        let via_router = self
-                            .router_venues
-                            .contains(&config.address);
-                        let component =
-                            build_component(&self.tokens, config, id, &token0, &token1, via_router);
-                        new_pairs.insert(id_string.clone(), component.clone());
-                        component
-                    });
-
-                let state = PriceLevelStreamState::new(
-                    token0,
-                    token1,
-                    quotes_0_to_1,
-                    quotes_1_to_0,
-                    config.gas_cost.clone(),
-                );
-
-                states.insert(id_string.clone(), Box::new(state));
-                self.components
-                    .insert(id_string, component);
-            }
-        }
-
-        // Every re-emitted pair was moved back into `self.components` above — whatever remains
-        // is gone: the pair, or its whole venue, is no longer streamed.
-        let removed_pairs = previous;
-
-        if states.is_empty() && new_pairs.is_empty() && removed_pairs.is_empty() {
-            return None;
-        }
-
-        Some(
-            // Quotes target the block currently being built, hence partial. Sync states stay
-            // empty (like the RFQ path) because no full block header is available.
-            Update::new(message.block_number, states, new_pairs)
-                .set_is_partial(true)
-                .set_removed_pairs(removed_pairs),
-        )
-    }
-}
-
-fn build_component(
-    tokens: &HashMap<Bytes, Token>,
-    config: &PriceLevelStreamConfig,
-    id: Bytes,
-    token0: &Bytes,
-    token1: &Bytes,
-    via_router: bool,
-) -> ProtocolComponent {
-    let protocol_system =
-        if via_router { config.fallback_protocol_system() } else { config.protocol_system() };
-    ProtocolComponent::new(
-        id,
-        protocol_system.clone(),
-        protocol_system,
-        // Titan builds Ethereum L1 blocks; the stream carries no other chains.
-        Chain::Ethereum,
-        vec![tokens[token0].clone(), tokens[token1].clone()],
-        vec![config.address.clone()],
-        HashMap::from([(PAMM_ADDRESS_ATTRIBUTE.to_string(), config.address.clone())]),
-        Bytes::default(),
-        Utc::now().naive_utc(),
-    )
-}
-
-/// The component identity of a (pAMM, pair) combination: `pamm ++ token0 ++ token1`.
-fn component_id(pamm: &Bytes, token0: &Bytes, token1: &Bytes) -> Bytes {
-    Bytes::from([pamm.as_ref(), token0.as_ref(), token1.as_ref()].concat())
-}
-
 #[cfg(test)]
 mod tests {
     use std::str::FromStr;
 
-    use num_bigint::BigUint;
-
-    use super::*;
-
-    const PAMM: &str = "0x5979458912f80b96d30d4220af8e2e4925a33320";
-    const WBTC: &str = "0x2260fac5e5542a773aa44fbcfedf7c193bc2c599";
-    const USDC: &str = "0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48";
-    const WETH: &str = "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2";
-
-    fn token(address: &str, symbol: &str, decimals: u32) -> Token {
-        Token::new(
-            &Bytes::from_str(address).unwrap(),
-            symbol,
-            decimals,
-            0,
-            &[Some(10_000)],
-            Chain::Ethereum,
-            100,
-        )
-    }
-
-    fn tokens() -> HashMap<Bytes, Token> {
-        [token(WBTC, "WBTC", 8), token(USDC, "USDC", 6), token(WETH, "WETH", 18)]
-            .into_iter()
-            .map(|token| (token.address.clone(), token))
-            .collect()
-    }
-
-    fn tracker() -> SnapshotTracker {
-        let config = PriceLevelStreamConfig::new(
-            "fermiswap",
-            Bytes::from_str(PAMM).unwrap(),
-            BigUint::from(120_000u64),
-        );
-        SnapshotTracker::new(
-            HashMap::from([(config.address.clone(), config)]),
-            HashSet::new(),
-            tokens(),
-            false,
-            BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST),
-            HashSet::new(),
-        )
-    }
-
-    fn level(amount_in: u64, amount_out: u64) -> TitanPriceLevel {
-        TitanPriceLevel {
-            amount_in: BigUint::from(amount_in),
-            amount_out: BigUint::from(amount_out),
-        }
-    }
-
-    fn pair_levels(
-        token_in: &str,
-        token_out: &str,
-        order_book: Vec<TitanPriceLevel>,
-    ) -> TitanPairLevels {
-        TitanPairLevels {
-            token_in: Bytes::from_str(token_in).unwrap(),
-            token_out: Bytes::from_str(token_out).unwrap(),
-            order_book,
-        }
-    }
-
-    fn message(block_number: u64, pairs: Vec<TitanPairLevels>) -> TitanPriceLevelMessage {
-        TitanPriceLevelMessage {
-            block_number,
-            timestamp: 0,
-            pamms: vec![TitanPammLevels { pamm: Bytes::from_str(PAMM).unwrap(), pairs }],
-        }
-    }
-
-    fn wbtc_usdc_pairs() -> Vec<TitanPairLevels> {
-        vec![
-            pair_levels(WBTC, USDC, vec![level(100_000_000, 100_000_000_000)]),
-            pair_levels(USDC, WBTC, vec![level(100_000_000_000, 99_000_000)]),
-        ]
-    }
-
-    fn expected_id() -> String {
-        // pamm ++ token0 ++ token1 with WBTC < USDC.
-        format!("{PAMM}{}{}", &WBTC[2..], &USDC[2..])
-    }
-
-    #[test]
-    fn first_snapshot_emits_new_pair_with_both_directions() {
-        let mut tracker = tracker();
-        let Update {
-            block_number_or_timestamp,
-            is_partial,
-            sync_states,
-            states,
-            new_pairs,
-            removed_pairs,
-        } = tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        assert_eq!(block_number_or_timestamp, 100);
-        assert!(is_partial);
-        assert!(sync_states.is_empty());
-        assert!(removed_pairs.is_empty());
-
-        let id = expected_id();
-        let component = &new_pairs[&id];
-        assert_eq!(component.protocol_system, "pricelevelstream:fermiswap");
-        assert_eq!(
-            component.static_attributes[PAMM_ADDRESS_ATTRIBUTE],
-            Bytes::from_str(PAMM).unwrap()
-        );
-
-        let PriceLevelStreamState {
-            token0,
-            token1,
-            quotes_0_to_1,
-            quotes_1_to_0,
-            gas_cost,
-            quotable_until: _,
-        } = states[&id]
-            .as_any()
-            .downcast_ref::<PriceLevelStreamState>()
-            .expect("price level state");
-        assert_eq!(token0, &Bytes::from_str(WBTC).unwrap());
-        assert_eq!(token1, &Bytes::from_str(USDC).unwrap());
-        assert_eq!(quotes_0_to_1.len(), 1);
-        assert_eq!(quotes_1_to_0.len(), 1);
-        assert_eq!(quotes_0_to_1[0].amount_in, BigUint::from(100_000_000u64));
-        assert_eq!(gas_cost, &BigUint::from(120_000u64));
-    }
-
-    #[test]
-    fn repeated_snapshot_is_not_a_new_pair() {
-        let mut tracker = tracker();
-        tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-        let update = tracker
-            .process(message(101, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        assert!(update.new_pairs.is_empty());
-        assert!(update.removed_pairs.is_empty());
-        assert!(update
-            .states
-            .contains_key(&expected_id()));
-    }
-
-    #[test]
-    fn dropped_pair_is_removed() {
-        let mut tracker = tracker();
-        tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-        let weth_usdc =
-            vec![pair_levels(WETH, USDC, vec![level(1_000_000_000_000_000_000, 3_000_000_000)])];
-        let update = tracker
-            .process(message(101, weth_usdc))
-            .expect("update expected");
-
-        assert_eq!(update.removed_pairs.len(), 1);
-        assert!(update
-            .removed_pairs
-            .contains_key(&expected_id()));
-        assert_eq!(update.new_pairs.len(), 1);
-        assert_eq!(update.states.len(), 1);
-    }
-
-    #[test]
-    fn out_of_order_frame_is_skipped() {
-        let mut tracker = tracker();
-        tracker
-            .process(message(101, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        // A frame for an older block is stale: no update, and the caches stay untouched even
-        // though the frame's snapshot differs completely.
-        let stale =
-            vec![pair_levels(WETH, USDC, vec![level(1_000_000_000_000_000_000, 3_000_000_000)])];
-        assert!(tracker
-            .process(message(100, stale))
-            .is_none());
-
-        // The next current frame diffs against the pre-stale state: nothing was added or
-        // removed in between.
-        let update = tracker
-            .process(message(102, wbtc_usdc_pairs()))
-            .expect("update expected");
-        assert!(update.new_pairs.is_empty());
-        assert!(update.removed_pairs.is_empty());
-    }
-
-    #[test]
-    fn vanished_pamm_has_its_pairs_removed() {
-        let mut tracker = tracker();
-        tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        // The next frame no longer contains the pAMM at all: a complete snapshot without a
-        // venue means the venue is gone, pairs and all.
-        let update = tracker
-            .process(TitanPriceLevelMessage { block_number: 101, timestamp: 0, pamms: vec![] })
-            .expect("update expected");
-        assert!(update.states.is_empty());
-        assert!(update.new_pairs.is_empty());
-        assert_eq!(update.removed_pairs.len(), 1);
-        assert!(update
-            .removed_pairs
-            .contains_key(&expected_id()));
-
-        // Nothing served and nothing changed: no update.
-        assert!(tracker
-            .process(TitanPriceLevelMessage { block_number: 102, timestamp: 0, pamms: vec![] })
-            .is_none());
-
-        // A venue that reappears is a new pair again.
-        let update = tracker
-            .process(message(103, wbtc_usdc_pairs()))
-            .expect("update expected");
-        assert!(update
-            .new_pairs
-            .contains_key(&expected_id()));
-    }
-
-    #[test]
-    fn unregistered_pamm_produces_no_update_without_auto_detection() {
-        let mut tracker = SnapshotTracker::new(
-            HashMap::new(),
-            HashSet::new(),
-            tokens(),
-            false,
-            BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST),
-            HashSet::new(),
-        );
-        assert!(tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .is_none());
-    }
-
-    #[test]
-    fn denied_pamm_is_not_auto_detected() {
-        let denied = HashSet::from([Bytes::from_str(PAMM).unwrap()]);
-        let mut tracker = SnapshotTracker::new(
-            HashMap::new(),
-            denied,
-            tokens(),
-            true,
-            BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST),
-            HashSet::new(),
-        );
-        assert!(tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .is_none());
-    }
+    use super::{
+        super::{config::default_denied_pamms, test_support::PAMM},
+        *,
+    };
 
     #[test]
     fn explicit_add_and_deny_are_last_wins() {
@@ -819,56 +371,6 @@ mod tests {
     }
 
     #[test]
-    fn auto_detected_pamm_is_served_under_its_address() {
-        let mut tracker = SnapshotTracker::new(
-            HashMap::new(),
-            HashSet::new(),
-            tokens(),
-            true,
-            BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST),
-            HashSet::new(),
-        );
-        let update = tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        let component = &update.new_pairs[&expected_id()];
-        assert_eq!(component.protocol_system, format!("pricelevelstream:{PAMM}"));
-        let state = update.states[&expected_id()]
-            .as_any()
-            .downcast_ref::<PriceLevelStreamState>()
-            .expect("price level state");
-        assert_eq!(state.gas_cost, BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST));
-
-        // The synthesized config is cached: the next snapshot is not a new pair again.
-        let update = tracker
-            .process(message(101, wbtc_usdc_pairs()))
-            .expect("update expected");
-        assert!(update.new_pairs.is_empty());
-    }
-
-    #[test]
-    fn auto_detected_gas_cost_override_applies() {
-        let mut tracker = SnapshotTracker::new(
-            HashMap::new(),
-            HashSet::new(),
-            tokens(),
-            true,
-            BigUint::from(42_000u64),
-            HashSet::new(),
-        );
-        let update = tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        let state = update.states[&expected_id()]
-            .as_any()
-            .downcast_ref::<PriceLevelStreamState>()
-            .expect("price level state");
-        assert_eq!(state.gas_cost, BigUint::from(42_000u64));
-    }
-
-    #[test]
     fn with_known_pamms_registers_known_venues() {
         // PAMM is the FermiSwap router, one of the default venues.
         let fermiswap_router = Bytes::from_str(PAMM).unwrap();
@@ -903,84 +405,6 @@ mod tests {
             assert_eq!(builder.registry[&fermiswap_router].protocol, "custom");
             assert_eq!(builder.registry[&fermiswap_router].gas_cost, BigUint::from(1u64));
         }
-    }
-
-    /// A venue on the router's whitelist is emitted under `propammfallback:{name}`, so its swaps
-    /// execute through Titan's PropAMMRouter; identity and attributes stay the same.
-    #[test]
-    fn whitelisted_venue_is_served_under_the_fallback_family() {
-        let config = PriceLevelStreamConfig::new(
-            "fermiswap",
-            Bytes::from_str(PAMM).unwrap(),
-            BigUint::from(120_000u64),
-        );
-        let mut tracker = SnapshotTracker::new(
-            HashMap::from([(config.address.clone(), config)]),
-            HashSet::new(),
-            tokens(),
-            false,
-            BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST),
-            HashSet::from([Bytes::from_str(PAMM).unwrap()]),
-        );
-
-        let update = tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        let component = &update.new_pairs[&expected_id()];
-        assert_eq!(component.protocol_system, "propammfallback:fermiswap");
-        assert_eq!(
-            component.static_attributes[PAMM_ADDRESS_ATTRIBUTE],
-            Bytes::from_str(PAMM).unwrap()
-        );
-    }
-
-    /// The whitelist check is by address, so it also covers auto-detected, address-named venues.
-    #[test]
-    fn auto_detected_whitelisted_venue_is_served_under_the_fallback_family() {
-        let mut tracker = SnapshotTracker::new(
-            HashMap::new(),
-            HashSet::new(),
-            tokens(),
-            true,
-            BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST),
-            HashSet::from([Bytes::from_str(PAMM).unwrap()]),
-        );
-
-        let update = tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        let component = &update.new_pairs[&expected_id()];
-        assert_eq!(component.protocol_system, format!("propammfallback:{PAMM}"));
-    }
-
-    /// A venue absent from the whitelist keeps the direct `pricelevelstream:{name}` family — the
-    /// router reverts `UnknownVenue` for it, which would send every swap to the Uniswap V3
-    /// fallback.
-    #[test]
-    fn unwhitelisted_venue_keeps_the_direct_family() {
-        let other_venue =
-            Bytes::from_str("0x71e790dd841c8a9061487cb3e78c288e75ce0b3d").expect("valid address");
-        let config = PriceLevelStreamConfig::new(
-            "fermiswap",
-            Bytes::from_str(PAMM).unwrap(),
-            BigUint::from(120_000u64),
-        );
-        let mut tracker = SnapshotTracker::new(
-            HashMap::from([(config.address.clone(), config)]),
-            HashSet::new(),
-            tokens(),
-            false,
-            BigUint::from(DEFAULT_AUTO_DETECTED_GAS_COST),
-            HashSet::from([other_venue]),
-        );
-
-        let update = tracker
-            .process(message(100, wbtc_usdc_pairs()))
-            .expect("update expected");
-
-        assert_eq!(update.new_pairs[&expected_id()].protocol_system, "pricelevelstream:fermiswap");
     }
 
     /// The PropAMMRouter path is the default; `without_fallback_router` is the way off it.
@@ -1021,18 +445,5 @@ mod tests {
 
         assert_eq!(format!("{PRICE_LEVEL_STREAM_FAMILY}:"), PRICE_LEVEL_STREAM_PREFIX);
         assert_eq!(format!("{PROPAMM_FALLBACK_FAMILY}:"), PROPAMM_FALLBACK_PREFIX);
-    }
-
-    #[test]
-    fn unknown_tokens_are_skipped() {
-        let mut tracker = tracker();
-        let unknown = vec![pair_levels(
-            "0x1111111111111111111111111111111111111111",
-            USDC,
-            vec![level(1, 1)],
-        )];
-        assert!(tracker
-            .process(message(100, unknown))
-            .is_none());
     }
 }
